requirements.txt
streamlit>=1.23.0
pandas>=1.5.3
databricks-sql-connector>=2.9.0
databricks-sdk>=0.19.0
pyarrow>=12.0.0
streamlit-searchbox>=0.1.3
